@@ -32,6 +32,7 @@
     "react": ">=15.0 || 0.14.x"
   },
   "devDependencies": {
+    "create-react-class": "^15.6.3",
     "eslint": "4.16.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-config-prettier": "2.9.0",
@@ -40,10 +41,10 @@
     "eslint-plugin-prettier": "2.5.0",
     "eslint-plugin-react": "^7.5.1",
     "lodash-webpack-plugin": "^0.8.1",
-    "nwb": "0.9.x",
+    "nwb": "^0.20.x",
     "prettier": "1.10.2",
-    "react": "0.14.x",
-    "react-dom": "0.14.x"
+    "react": "0.16.2",
+    "react-dom": "0.16.2"
   },
   "authors": [
     "Michael Schneider <mischneider1@gmail.com> (https://github.com/maicki)",
